requirements.txt
python-dotenv
pandas
sklearn
numpy
matplotlib
vaderSentiment
datetime
nltk
scipy
statsmodels
seaborn
apt